requirements.txt
telebot
Pillow
pytesseract